Add ssl default settings in http context. (#13)
The ssl settings that are done in the http context used as default for
all server within the http conext. So with this new option we can manage
ssl settings in a single place. The settings can be set via pillar. If
needed the configuration values can be override in each server.

Co-authored-by: Daniel Goeke <daniel.goeke@ioki.com>
Co-authored-by: Chris Aumann <me@chr4.org>