Use different ways to set ciphers for TLS 1.3
Openssl have a different way to set ciphers for TLS 1.3. If you using
ssl_ciphers with only TLS 1.3 nginx will fail to start. You need to use
"ssl_conf_command Ciphersuites" command in nginx.

I implement it that way that I check in the template if tls version is
set to 1.3 if so use the new syntax.

Links:
https://forum.nginx.org/read.php?11,287698
https://github.com/mozilla/ssl-config-generator/issues/124
https://wiki.openssl.org/index.php/TLS1.3